separate base quantum tape from quantum tape
